pyproject.toml
[tool.poetry]
name = "desmos-student-image"
version = "0.1.0"
description = ""
authors = ["Nihaal Sangha <me@niha.al>"]
license = "GPL-3.0-only"

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^8.1.0"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
